Adds about screen and fixed disk size

+++ firmware/src/Files/FlashLittleFS.h
@@ -14,7 +14,7 @@ public:
   bool isMounted() {
     return _isMounted;
   }
-  bool getSpace(size_t &total, size_t &used);
+  bool getSpace(uint64_t &total, uint64_t &used);
   const char *mountPoint() {
     return m_mountPoint.c_str();
   }
